Divided one test file into several
